fix code smells

+++ src/main/java/ru/r2cloud/satellite/decoder/GmskDemodulator.java
@@ -25,7 +25,7 @@ public class GmskDemodulator implements ByteInput {
 		FLLBandEdge fll = new FLLBandEdge(agc, samplesPerSymbol, 0.35f, 100, 0.06f);
 		LowPassFilterComplex lpf = new LowPassFilterComplex(fll, 1.0, bandwidth / 2, 600, Window.WIN_HAMMING, 6.76);
 		QuadratureDemodulation qd = new QuadratureDemodulation(lpf, 1.0f);
-		LowPassFilter lpf2 = new LowPassFilter(qd, 1.0, baudRate / 2, 2000, Window.WIN_HAMMING, 6.76);
+		LowPassFilter lpf2 = new LowPassFilter(qd, 1.0, (double) baudRate / 2, 2000, Window.WIN_HAMMING, 6.76);
 		ClockRecoveryMM clockRecovery = new ClockRecoveryMM(lpf2, samplesPerSymbol, (float) (0.25 * gainMu * gainMu), 0.5f, gainMu, 0.005f);
 		Rail rail = new Rail(clockRecovery, -1.0f, 1.0f);
 		this.source = new FloatToChar(rail, 127.0f);
